Configure adminish with a couchish store config.
All the configuration about the model types adminish handles is in
the couchish store's config so it might as well just get it from there.

In addition, there was a rather ugly reliance on a 'couchish' key in the
environ. Instead, the adminish config now includes a store factory that
is used to retrieve the couchish store related to the current request.